group: UIStackView, Double extension 파일 분리
